Downloads/auto-spare-parts/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
pillow==10.1.0
psycopg2-binary==2.9.9
django-parler==2.3.0
django-parler-rest==2.2
django-filter==23.3
python-dotenv==1.0.0
django-storages==1.14.2
django-cleanup==8.0.0
django-rest-auth==0.9.5
django-allauth==0.57.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.25
django-countries==7.5.1
django-mptt==0.14.0
django-taggit==5.0.1
django-ckeditor==6.7.0
django-imagekit==4.1.0
django-simple-history==3.4.0
django-admin-honeypot==1.1.0
django-ratelimit==4.1.0
django-cacheops==7.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
